Hackathon final result - first time coding
We received the newbie award for this because we had the best project
of any of the high schoolers there. I worked with two other girls who
had also done little to no coding before. They made the calculator, I
made the user interface and combined our code. It’s meant to be a
homework helper that works like a calculator for certain scientific
equations.